requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
click==8.3.0
discord==2.3.2
discord.py==2.6.3
Flask==3.1.2
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mcrcon==0.7.0
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
pytz==2025.2
typing_extensions==4.15.0
Werkzeug==3.1.3
yarl==1.20.1